+++ fiddles/angular2/fiddle-0026-Nvd3Component/src/meta.js
@@ -1,0 +1,9 @@
+var metadata = {
+    fiddleHeader: 'Angular 2 - TreeView',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0025-TreeView',
+        data: "data.json"
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};
+//# sourceMappingURL=meta.js.map