+++ fiddles/d3/fiddle-0002-SingleSelection/app.js
@@ -1,0 +1,5 @@
+(function (app, $, undefined) {
+    "use strict";
+    d3.select("p#target")
+        .text("Hello world!");
+})(window.app = window.app || {}, d3)