+++ fiddles/d3/fiddle-0003-MultipleSelect/app.js
@@ -1,0 +1,5 @@
+(function (app, $, undefined) {
+    "use strict";
+    d3.selectAll("div")
+        .attr("class", "red box");
+})(window.app = window.app || {}, d3)