+++ fiddles/d3/fiddle-0004-SelectionIteration/app.js
@@ -1,0 +1,9 @@
+(function (app, $, undefined) {
+    "use strict";
+    d3.selectAll("div")
+        .attr("class", "red box")
+        .each(function (d, i) {
+            d3.select(this).append("h1").text(i);
+        });
+})(window.app = window.app || {}, d3)
+