+++ fiddles/d3/fiddle-0006-FunctionChain/app.js
@@ -1,0 +1,21 @@
+(function (app, $, undefined) {
+    "use strict";
+
+    d3.select("body")
+        .append("section")
+        .attr("id", "section1")
+        .append("div")
+        .attr("class", "blue box")
+        .append("p") // <-- F
+        .text("dynamic blue box");
+
+    d3.select("body")
+        .append("section")
+        .attr("id", "section2")
+        .append("div")
+        .attr("class", "red box")
+        .append("p")
+        .text("dynamic red box");
+
+
+})(window.app = window.app || {}, d3)